More Q# tests
