--- qir/qat/Validator/Tests/Unit/QSharpTypeQirs.cpp
@@ -1,125 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#include "Generators/DefaultProfileGenerator.hpp"
-#include "Llvm/Llvm.hpp"
-#include "Rules/Factory.hpp"
-#include "TestTools/IrManipulationTestHelper.hpp"
-#include "gtest/gtest.h"
-
-#include <functional>
-using namespace microsoft::quantum;
-
-namespace {
-using IrManipulationTestHelperPtr = std::shared_ptr<IrManipulationTestHelper>;
-IrManipulationTestHelperPtr newIrManip(std::string const &script)
-{
-  IrManipulationTestHelperPtr ir_manip = std::make_shared<IrManipulationTestHelper>();
-
-  ir_manip->declareOpaque("Qubit");
-  ir_manip->declareOpaque("Result");
-  ir_manip->declareOpaque("Array");
-  ir_manip->declareOpaque("Tuple");
-  ir_manip->declareOpaque("Range");
-  ir_manip->declareOpaque("Callable");
-  ir_manip->declareOpaque("String");
-
-  ir_manip->declareFunction("void @__quantum__qis__h__body(%Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__x__body(%Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__s__body(%Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__cx__body(%Qubit*, %Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__h__body(%Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__rz__body(double, %Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__t__body(%Qubit*)");
-  ir_manip->declareFunction("void @__quantum__qis__z__body(%Qubit*)");
-
-  ir_manip->declareFunction("void @__quantum__qis__mz__body(%Qubit*, %Result*)");
-  ir_manip->declareFunction("void @__quantum__qis__reset__body(%Qubit*)");
-  ir_manip->declareFunction("i1 @__quantum__qis__read_result__body(%Result*)");
-  ir_manip->declareFunction("void @__quantum__qis__cnot__body(%Qubit*, %Qubit*)");
-
-  if (!ir_manip->fromBodyString(script))
-  {
-    llvm::outs() << ir_manip->generateScript(script) << "\n\n";
-    llvm::outs() << ir_manip->getErrorMessage() << "\n";
-    exit(-1);
-  }
-  return ir_manip;
-}
-
-void expectSuccess(String const &profile_name, String const &script)
-{
-  auto ir_manip = newIrManip(script);
-
-  auto profile_generator = std::make_shared<DefaultProfileGenerator>();
-
-  ConfigurationManager &configuration_manager = profile_generator->configurationManager();
-  configuration_manager.addConfig<FactoryConfiguration>();
-  configuration_manager.addConfig<ValidationPassConfiguration>(
-      "validation-configuration", ValidationPassConfiguration::fromProfileName(profile_name));
-
-  EXPECT_TRUE(ir_manip->validateProfile(profile_generator, profile_name));
-}
-
-}  // namespace
-
-TEST(QSharpPositive, TeleportChain)
-{
-  expectSuccess("base", R"script(
-  tail call void @__quantum__qis__h__body(%Qubit* null)
-  tail call void @__quantum__qis__cnot__body(%Qubit* null, %Qubit* nonnull inttoptr (i64 1 to %Qubit*))
-  tail call void @__quantum__qis__h__body(%Qubit* nonnull inttoptr (i64 2 to %Qubit*))
-  tail call void @__quantum__qis__cnot__body(%Qubit* nonnull inttoptr (i64 2 to %Qubit*), %Qubit* nonnull inttoptr (i64 4 to %Qubit*))
-  tail call void @__quantum__qis__h__body(%Qubit* nonnull inttoptr (i64 3 to %Qubit*))
-  tail call void @__quantum__qis__cnot__body(%Qubit* nonnull inttoptr (i64 3 to %Qubit*), %Qubit* nonnull inttoptr (i64 5 to %Qubit*))
-  tail call void @__quantum__qis__cnot__body(%Qubit* nonnull inttoptr (i64 1 to %Qubit*), %Qubit* nonnull inttoptr (i64 2 to %Qubit*))
-  tail call void @__quantum__qis__h__body(%Qubit* nonnull inttoptr (i64 1 to %Qubit*))
-  tail call void @__quantum__qis__mz__body(%Qubit* nonnull inttoptr (i64 1 to %Qubit*), %Result* null)
-  tail call void @__quantum__qis__reset__body(%Qubit* nonnull inttoptr (i64 1 to %Qubit*))
-  %0 = tail call i1 @__quantum__qis__read_result__body(%Result* null)
-  br i1 %0, label %quantum, label %quantum27
-
-quantum:                                          ; preds = %entry
-  tail call void @__quantum__qis__z__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*))
-  br label %quantum27
-
-quantum27:                                        ; preds = %quantum, %entry
-  tail call void @__quantum__qis__mz__body(%Qubit* nonnull inttoptr (i64 2 to %Qubit*), %Result* nonnull inttoptr (i64 1 to %Result*))
-  tail call void @__quantum__qis__reset__body(%Qubit* nonnull inttoptr (i64 2 to %Qubit*))
-  %1 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 1 to %Result*))
-  br i1 %1, label %quantum2, label %quantum32
-
-quantum2:                                         ; preds = %quantum27
-  tail call void @__quantum__qis__x__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*))
-  br label %quantum32
-
-quantum32:                                        ; preds = %quantum2, %quantum27
-  tail call void @__quantum__qis__cnot__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*), %Qubit* nonnull inttoptr (i64 3 to %Qubit*))
-  tail call void @__quantum__qis__h__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*))
-  tail call void @__quantum__qis__mz__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*), %Result* nonnull inttoptr (i64 2 to %Result*))
-  tail call void @__quantum__qis__reset__body(%Qubit* nonnull inttoptr (i64 4 to %Qubit*))
-  %2 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 2 to %Result*))
-  br i1 %2, label %quantum7, label %quantum37
-
-quantum7:                                         ; preds = %quantum32
-  tail call void @__quantum__qis__z__body(%Qubit* nonnull inttoptr (i64 5 to %Qubit*))
-  br label %quantum37
-
-quantum37:                                        ; preds = %quantum7, %quantum32
-  tail call void @__quantum__qis__mz__body(%Qubit* nonnull inttoptr (i64 3 to %Qubit*), %Result* nonnull inttoptr (i64 3 to %Result*))
-  tail call void @__quantum__qis__reset__body(%Qubit* nonnull inttoptr (i64 3 to %Qubit*))
-  %3 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 3 to %Result*))
-  br i1 %3, label %quantum12, label %quantum17
-
-quantum12:                                        ; preds = %quantum37
-  tail call void @__quantum__qis__x__body(%Qubit* nonnull inttoptr (i64 5 to %Qubit*))
-  br label %quantum17
-
-quantum17:                                        ; preds = %quantum12, %quantum37
-  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* nonnull inttoptr (i64 4 to %Result*))
-  tail call void @__quantum__qis__reset__body(%Qubit* null)
-  tail call void @__quantum__qis__mz__body(%Qubit* nonnull inttoptr (i64 5 to %Qubit*), %Result* nonnull inttoptr (i64 5 to %Result*))
-  tail call void @__quantum__qis__reset__body(%Qubit* nonnull inttoptr (i64 5 to %Qubit*))
-
-  )script");
-}
