+++ qir/qsharp/ConstArray/Makefile
@@ -1,7 +1,7 @@
 all: QSharpVersion/qir/Example.ll
 	${QAT_BINARY}  QSharpVersion/qir/Example.ll ${QIR_STDLIB} -O3 -S --apply > combined.ll 
 	opt -O3 -S combined.ll > optimised.ll
-#	${QAT_BINARY} -O3 -S --apply --validate --save-validation-report validation.json --profile base optimised.ll > /dev/null
+	${QAT_BINARY} -S --validate --save-validation-report validation.json  --profile base optimised.ll || echo "failed!"
 	@echo "Validation for ConstArray finished"
