+++ qir/qsharp/LoopRecursion/Makefile
@@ -3,7 +3,7 @@ all: qat-qsharp
 qat-qsharp: QSharpVersion/qir/Example.ll
 	${QAT_BINARY} -O3 --apply -S --verify-module  --profile base QSharpVersion/qir/Example.ll > combined.ll
 	opt -O3 -S combined.ll > optimised.ll
-	${QAT_BINARY} -O3 -S --apply --validate --save-validation-report validation.json --profile base optimised.ll > /dev/null
+	${QAT_BINARY} -S --validate --save-validation-report validation.json  --profile base optimised.ll || echo "failed!"
 	@echo "Validation for LoopRecursion finished"
