+++ qir/qsharp/OptimisationIssue/validation.json
@@ -1,0 +1,79 @@
+[
+  {
+    "type": "error",
+    "message": "'store' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  store %Qubit* null, %Qubit** %2, align 8",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'load' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %77 = load %Qubit*, %Qubit** %42, align 8",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'bitcast' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %2 = bitcast i8* %1 to %Qubit**",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__array_get_element_ptr_1d' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %1 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %0, i64 0)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__array_create_1d' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %0 = tail call %Array* @__quantum__rt__array_create_1d(i32 8, i64 20)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type 'Array*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %1 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %0, i64 0)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Fatal error: QIR is not valid within the defined profile",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %1 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %0, i64 0)",
+      "frontend_hint": ""
+    }
+  }
+]
