+++ qir/qsharp/SimpleCallable/validation.json
@@ -1,0 +1,266 @@
+[
+  {
+    "type": "error",
+    "message": "'zext' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %31 = zext i1 %30 to i64",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'icmp' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %.not2.i.i = icmp slt i64 %22, 0",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'store' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  store i64 2, i64* %1, align 4",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'add' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %22 = add i64 %21, -1",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'load' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %element.i.i = load i64, i64* %25, align 4",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'bitcast' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %1 = bitcast i8* %0 to i64*",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__capture_update_reference_count' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__capture_update_reference_count(%Callable* %20, i32 -1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__tuple_update_reference_count' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__tuple_update_reference_count(%Tuple* %26, i32 -1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__callable_invoke' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__callable_invoke(%Callable* %20, %Tuple* %26, %Tuple* %28)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__tuple_create' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %26 = tail call %Tuple* @__quantum__rt__tuple_create(i64 8)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__callable_update_reference_count' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__callable_update_reference_count(%Callable* %20, i32 -1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__array_get_size_1d' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %21 = tail call i64 @__quantum__rt__array_get_size_1d(%Array* %x.i)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__capture_update_alias_count' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__capture_update_alias_count(%Callable* %20, i32 1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__callable_create' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %20 = tail call %Callable* @__quantum__rt__callable_create([4 x void (%Tuple*, %Tuple*, %Tuple*)*]* @Microsoft__Quantum__Samples__T__FunctionTable, [2 x void (%Tuple*, i32)*]* null, %Tuple* null)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__array_get_element_ptr_1d' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %0 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %x.i, i64 0)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__callable_update_alias_count' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__callable_update_alias_count(%Callable* %20, i32 1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "'__quantum__rt__array_create_1d' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %x.i = tail call %Array* @__quantum__rt__array_create_1d(i32 8, i64 10)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type 'i1*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %30 = load i1, i1* %29, align 1",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type 'Tuple*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %20 = tail call %Callable* @__quantum__rt__callable_create([4 x void (%Tuple*, %Tuple*, %Tuple*)*]* @Microsoft__Quantum__Samples__T__FunctionTable, [2 x void (%Tuple*, i32)*]* null, %Tuple* null)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type '[2 x void (%Tuple*, i32)*]*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %20 = tail call %Callable* @__quantum__rt__callable_create([4 x void (%Tuple*, %Tuple*, %Tuple*)*]* @Microsoft__Quantum__Samples__T__FunctionTable, [2 x void (%Tuple*, i32)*]* null, %Tuple* null)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type '[4 x void (%Tuple*, %Tuple*, %Tuple*)*]*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %20 = tail call %Callable* @__quantum__rt__callable_create([4 x void (%Tuple*, %Tuple*, %Tuple*)*]* @Microsoft__Quantum__Samples__T__FunctionTable, [2 x void (%Tuple*, i32)*]* null, %Tuple* null)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type 'Callable*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  tail call void @__quantum__rt__capture_update_alias_count(%Callable* %20, i32 1)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Type 'Array*' is not allowed for this profile.",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %0 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %x.i, i64 0)",
+      "frontend_hint": ""
+    }
+  },
+  {
+    "type": "error",
+    "message": "Fatal error: QIR is not valid within the defined profile",
+    "location": {
+      "filename": "",
+      "row": 0,
+      "col": 0,
+      "llvm_hint": "  %0 = tail call i8* @__quantum__rt__array_get_element_ptr_1d(%Array* %x.i, i64 0)",
+      "frontend_hint": ""
+    }
+  }
+]
