+++ qir/qsharp/SimpleLoop/combined.ll
@@ -1,0 +1,57 @@
+; ModuleID = 'qat-link'
+source_filename = "qat-link"
+
+%Qubit = type opaque
+%Result = type opaque
+
+define i64 @SimpleLoop__Main__Interop() local_unnamed_addr #0 {
+entry:
+  br label %load
+
+load:                                             ; preds = %entry
+  br label %quantum
+
+quantum:                                          ; preds = %load
+  tail call void @__quantum__qis__h__body(%Qubit* null)
+  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* null)
+  tail call void @__quantum__qis__reset__body(%Qubit* null)
+  tail call void @__quantum__qis__h__body(%Qubit* null)
+  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* nonnull inttoptr (i64 1 to %Result*))
+  tail call void @__quantum__qis__reset__body(%Qubit* null)
+  tail call void @__quantum__qis__h__body(%Qubit* null)
+  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* nonnull inttoptr (i64 2 to %Result*))
+  tail call void @__quantum__qis__reset__body(%Qubit* null)
+  tail call void @__quantum__qis__h__body(%Qubit* null)
+  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* nonnull inttoptr (i64 3 to %Result*))
+  tail call void @__quantum__qis__reset__body(%Qubit* null)
+  tail call void @__quantum__qis__h__body(%Qubit* null)
+  tail call void @__quantum__qis__mz__body(%Qubit* null, %Result* nonnull inttoptr (i64 4 to %Result*))
+  tail call void @__quantum__qis__reset__body(%Qubit* null)
+  br label %readout
+
+readout:                                          ; preds = %quantum
+  %0 = tail call i1 @__quantum__qis__read_result__body(%Result* null)
+  %1 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 1 to %Result*))
+  %2 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 2 to %Result*))
+  %3 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 3 to %Result*))
+  %4 = tail call i1 @__quantum__qis__read_result__body(%Result* nonnull inttoptr (i64 4 to %Result*))
+  br label %post-classical
+
+post-classical:                                   ; preds = %readout
+  %5 = select i1 %4, i64 1337, i64 122
+  br label %exit_quantum_grouping
+
+exit_quantum_grouping:                            ; preds = %post-classical
+  ret i64 %5
+}
+
+declare void @__quantum__qis__reset__body(%Qubit*) local_unnamed_addr
+
+declare void @__quantum__qis__h__body(%Qubit*) local_unnamed_addr
+
+declare void @__quantum__qis__mz__body(%Qubit*, %Result*) local_unnamed_addr
+
+declare i1 @__quantum__qis__read_result__body(%Result*) local_unnamed_addr
+
+attributes #0 = { "InteropFriendly" "requiredQubits"="1" "requiredResults"="5" }
+
